validator using Environment variable and docker compose bash scripts
Added an environment variable to the validator docker file and creation of directories for configuration files,keys,data & logs. Docker compose file now uses a single bash script that checks if keys are already created or not before generating keys and starting validator connections.  (this is all in sawtooth-coffeechain-tp-testing.yaml)